Muda os nomes dos módulos
